[refactor] int -> Long파싱 #49

+++ src/main/java/_v3r/project/history/controller/UnrealHistoryController.java
@@ -26,8 +26,8 @@ public class UnrealHistoryController {
     @PostMapping("/{chatId}")
     @Operation(summary = "언리얼 통신 후 데이터 저장 기능")
     public CustomApiResponse<Void> receiveUnrealData(
-            @RequestHeader("user-no") Long userId,
-            @PathVariable Long chatId,
+            @RequestHeader("user-no") int userId,
+            @PathVariable int chatId,
             @RequestBody List<UnrealHistoryResponse> requestList
     ) {
         unrealHistoryService.saveAll(userId, chatId, requestList);
